requirements_working.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
array-api-compat==1.13.0
attrs==25.4.0
Authlib==1.6.6
backoff==2.2.1
bcrypt==5.0.0
beartype==0.22.9
biom_format==2.1.17
blinker==1.9.0
build==1.3.0
cachetools==5.5.0
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
chromadb==1.4.0
click==8.3.1
cloudpickle==3.1.2
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cyclopts==4.4.3
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
durationpy==0.10
email-validator==2.3.0
exceptiongroup==1.3.1
fakeredis==2.33.0
fastmcp==2.14.2
filelock==3.20.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
imbalanced-learn==0.12.3
importlib_metadata==8.7.1
importlib_resources==6.5.2
jaraco.classes==3.4.0
jaraco.context==6.0.2
jaraco.functools==4.4.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
kiwisolver==1.4.9
kubernetes==33.1.0
langchain==1.2.0
langchain-core==1.2.6
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
llvmlite==0.46.0
lupa==2.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mcp==1.25.0
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
narwhals==2.14.0
natsort==8.4.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.14.0
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging==24.2
pandas==2.3.3
pathable==0.4.4
pathlib==1.0.1
pathvalidate==3.3.1
patsy==1.0.2
pillow==11.1.0
platformdirs==4.5.1
plotly==6.5.0
posthog==5.4.0
prometheus_client==0.23.1
proto-plus==1.27.0
protobuf==5.29.5
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydocket==0.16.3
Pygments==2.19.2
PyJWT==2.10.1
pynndescent==0.5.13
pyparsing==3.3.1
pyperclip==1.11.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.36.2
regex==2025.11.3
reportlab==4.3.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-bio==0.7.1.post1
scikit-learn==1.5.1
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==3.3.1
setuptools==80.9.0
shap==0.50.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sse-starlette==3.1.2
starlette==0.50.0
statsmodels==0.14.6
streamlit==1.42.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.7
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
xgboost==2.1.0
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
